+++ src/apis/classDetail/queries/index.ts
@@ -1,9 +1,13 @@
-import { instance } from '@/apis/api';
-import { API_URL } from '@/apis/constants/apiURL';
+import { useQuery } from '@tanstack/react-query';
+import { getLessonDetail } from '@/apis/classDetail/axios';
+import { QUERY_KEYS } from '@/apis/constants/queryKey';
+import { LessonDetailGetResponse } from '@/types/myPageLessonDetailTypes';
 
-export const getLessonDetail = async (lessonId: number) => {
-  const url = `${API_URL.MEMBERS_LESSONS_DETAIL}/${lessonId}`;
+export const useGetLessonDetail = (lessonId: number) => {
+  return useQuery<LessonDetailGetResponse>({
+    queryKey: [QUERY_KEYS.MY_PAGE_LESSON_DETAIL, lessonId],
+    queryFn: () => getLessonDetail(lessonId),
 
-  const { data } = await instance.get(url);
-  return data;
+    enabled: lessonId !== undefined,
+  });
 };
