--- src/pages/instructor/classDetail/mocks/mockLessonDetail.ts
@@ -1,32 +1,0 @@
-export const LESSON_DETAIL = [
-  {
-    lessonId: 12345,
-    lessonName: '스트리트 댄스 기초',
-    lessonImageUrl: 'https://example.com/lesson-image.jpg',
-    lessonGenre: '힙합',
-    lessonLevel: '입문자',
-    lessonLocation: '서울특별시 강남구',
-    lessonStartDateTime: '2023-12-15T10:00:00',
-    lessonEndDateTime: '2023-12-15T12:00:00',
-    participants: [
-      {
-        reservationId: 1001,
-        bookerName: '홍길동',
-        bookerPhoneNumber: '010-1234-5678',
-        reservationDateTime: '2023-12-01T15:30:00',
-      },
-      {
-        reservationId: 1002,
-        bookerName: '이수진',
-        bookerPhoneNumber: '010-2345-6789',
-        reservationDateTime: '2023-12-01T15:45:00',
-      },
-      {
-        reservationId: 1003,
-        bookerName: '김민수',
-        bookerPhoneNumber: '010-3456-7890',
-        reservationDateTime: '2023-12-01T16:00:00',
-      },
-    ],
-  },
-];
